requirements.txt
aiogram[i18n]==3.4.1
APScheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
SQLAlchemy==2.0.29
user-agent==0.1.10
